dom/html/test/test_document-element-inserted.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: document-element-inserted</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<iframe id = 'media'>
</iframe>
<pre id="test">
<script class="testbody" type="text/javascript">

if (navigator.platform.startsWith("Win")) {
  SimpleTest.expectAssertions(0, 4);
}

SimpleTest.waitForExplicitFinish();
var loc;

var observe = function(doc){
    if (doc == media.contentDocument) {
        ok(media.contentDocument.location.toString().includes(loc),
          "The loaded media should be " + loc);
        next();
    }
}

var media = document.getElementById('media');
var tests = [
    "../../../media/test/short-video.ogv",
    "../../../media/test/sound.ogg",
    "../../content/test/image.png"
]

function next() {
    if (tests.length) {
        var t = tests.shift();
        loc = t.substring(t.indexOf("test"));
        media.setAttribute("src",t);
    }
    else {
        SpecialPowers.removeObserver(observe, "document-element-inserted");
        SimpleTest.finish();
    }
}

SpecialPowers.addObserver(observe, "document-element-inserted")
next();

</script>
</pre>
</body>
</html>
